+++ src/test/kotlin/lotto/util/BonusNumberValidatorTest.kt
@@ -1,5 +1,8 @@
-package lotto
+package lotto.util
 
+import lotto.Lotto
+import lotto.util.validator.bonusnumber.BonusNumberErrorType
+import lotto.util.validator.bonusnumber.BonusNumberValidator
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
